move bulk create studies class method
